APIs/presalytics.io/ooxml/0.1.0/python-aiohttp-client/tests/test_slides_group_element_types_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.problem_details import ProblemDetails
from openapi_server.models.slide_group_element_types import SlideGroupElementTypes


pytestmark = pytest.mark.asyncio

async def test_slides_groupelementtypes_get(client):
    """Test case for slides_groupelementtypes_get

    GroupElementTypes: List All Possible Types
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GroupElementTypes',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_slides_groupelementtypes_get_id(client):
    """Test case for slides_groupelementtypes_get_id

    GroupElementTypes: Get by Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GroupElementTypes/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_slides_groupelementtypes_typeid_get_type_id(client):
    """Test case for slides_groupelementtypes_typeid_get_type_id

    GroupElementTypes: Get By Type Id
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/ooxml-automation/Slides/GroupElementTypes/TypeId/{type_id}'.format(type_id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

